icon changes, forceToStop
a. Icons;
b. Crates folder if it is not existed;
c. Images naming: inputTitle_inputValue.png
d. Images naming: provide an option to overwrite the image name in
“Image Paramaters”
e. Images setting is optional to Aggregator;
f. ForceToStop: create an empty text file, and named “STOP.txt”, put it
to “C:\ladybug\DesignExplorer\”; it is checking if
“C:\ladybug\DesignExplorer\ STOP.txt” existed before move to next
iteration.